changed workload class to new representations, tests passing except mwem
